refactor: Moved cosmology module to separate directory
Moved the 'cosmology' module that is original one file
to its own directory.
